moved some functions that are not used regularly or maintained  in volcat.py to volcat_legacy.py
